Tests/LibWeb/Text/expected/HTML/document-open-cancels-parser-completion.txt

Replacement DOMContentLoaded event fired
Replacement iframe load event fired
